some final touches

+++ src/main/java/com/app/sortamine/algorithms/SortingStrategy.java
@@ -1,16 +1,15 @@
 package com.app.sortamine.algorithms;
 
 import com.app.sortamine.events.SortingEvent;
-import com.app.sortamine.events.SwapEvent;
 
 import java.util.List;
 
-public interface SortingStrategy { // TODO: color the sorted portion of the array
+public interface SortingStrategy {
     List<SortingEvent> generateSortHistory(int[] originalData);
 
     String getAlgorithmName();
 
-    int getComparisons();
+    long getComparisons();
 
-    int getInterchanges();
+    long getInterchanges();
 }
